feat: 일기장 인원 초대 및 수락기능 구현

+++ src/main/java/ac/mju/memoria/backend/domain/diarybook/repository/DiaryBookRepository.java
@@ -10,5 +10,5 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 
-public interface DiaryBookRepository extends JpaRepository<DiaryBook, Integer> {
+public interface DiaryBookRepository extends JpaRepository<DiaryBook, Long> {
 }
